Add promise support to MigrationDSL.prototype.addColumn
