+++ lib/migration-dsl.js
@@ -1,4 +1,5 @@
 var _  = require("lodash");
+const {callbackOrPromise} = require('./utils');
 
 //lets you attach further metadata to column definition
 //e.g. 'references product(id) on delete cascade'
@@ -62,7 +63,7 @@ MigrationDSL.prototype = {
     this.Dialect.createCollection(this.driver, collectionName, columns, keys, cb);
   },
 
-  addColumn: function (collectionName, options, cb) {
+  _addColumn: function (collectionName, options, cb) {
     var columnName = _.keys(options)[0]
     var column = this._createColumn(collectionName, columnName, options[columnName], this.Dialect, this.driver);
     this.Dialect.addCollectionColumn(this.driver, collectionName, column.value, null, cb);
@@ -120,6 +121,27 @@ MigrationDSL.prototype = {
   // , renameTable: function(oldCollectionName, newCollectionName, cb){
   //   this.Dialect.renameTable(this.driver, oldCollectionName, newCollectionName, cb);
   // }
-}
+};
+
+const callbackMethods = [
+  // 'createTable',
+  'addColumn',
+  // 'renameColumn',
+  // 'addIndex',
+  // 'dropIndex',
+  // 'dropColumn',
+  // 'dropTable',
+  // 'addPrimaryKey',
+  // 'dropPrimaryKey',
+  // 'addForeignKey',
+  // 'dropForeignKey',
+  // 'hasTable',
+  // 'getColumns',
+  // 'execQuery'
+];
+
+callbackMethods.forEach(methodName => {
+  MigrationDSL.prototype[methodName] = callbackOrPromise(MigrationDSL.prototype['_' + methodName]);
+});
 
 module.exports = MigrationDSL;
