+++ lib/migrator.js
@@ -8,6 +8,7 @@ const Promise = require('bluebird');
 
 var Migration    = require('./migration');
 var MigrationDsl = require('./migration-dsl');
+const {callbackOrPromise} = require('./utils');
 
 /**
  * Log a keyed message.
@@ -122,19 +123,6 @@ Migrator.prototype.loadModules = function() {
   });
 }
 
-// TODO: move this into a separate module and provide tests fot this...
-function callbackOrPromise(originalMethod) {
-  return function() {
-    const ctx = this;
-
-    const cb = _.last(arguments);
-    if(typeof cb === "function") {
-      return originalMethod.apply(ctx, arguments);
-    } else {
-      return Promise.promisify(originalMethod).apply(ctx, arguments);
-    }
-  }
-}
 /**
  * Perform a migration in the given `direction`.
  *
